product test and modified destroy function

+++ app/Http/Controllers/Backend/ProductsController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\Backend\ProductsRequest;
 use App\Http\Controllers\Controller;
 use App\Models\Backend\Product;
 use App\Models\Backend\Brand;
+use App\Models\Backend\OrderDetails;
 
 class ProductsController extends Controller
 {
@@ -113,9 +114,17 @@ class ProductsController extends Controller
      */
     public function destroy($id)
     {
-        $product = Product::findOrFail($id);
-        $product->delete();
+        $product = OrderDetails::where('product_id', $id)->first();
+        
+        if ($product === null) {
+            $product = Product::findOrFail($id);
+            $product->delete();
 
-        return back();
+            session()->flash('message', 'Product was deleted successfully!');
+        } else {
+            session()->flash('message', 'You can\'t deleted this product!');
+        }
+
+        return redirect('admin/products');
     }
 }
